siru-slm/requirements.txt
transformers>=4.40.0
peft>=0.10.0
datasets>=2.19.0
trl>=0.8.0
bitsandbytes>=0.43.0
accelerate>=0.29.0
torch>=2.2.0
vllm>=0.4.0
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.0
openai>=1.25.0
supabase>=2.4.0
python-dotenv>=1.0.0
httpx>=0.27.0
tiktoken>=0.7.0
rich>=13.7.0
replicate>=0.26.0
